requirements.txt
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
Deprecated==1.2.14
idna==3.4
pycparser==2.21
PyGithub==1.55
PyJWT==2.8.0
PyNaCl==1.5.0
requests==2.31.0
tomli==2.0.1
urllib3==2.0.7
wrapt==1.15.0
